+++ db/db_user.py
@@ -33,24 +33,28 @@ def get_user_by_username(db: Session, username: str):
       detail=f'User with username {username} not found')
   return user
 
-def update_user(db: Session, id: int, request: UserBase):
+def update_user(db: Session, id: int, request: UserBase,current_user):
   user = db.query(DbUser).filter(DbUser.id == id)
   if not user.first():
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
       detail=f'User with id {id} not found')
+  if current_user.id != id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='You are not allowed to update this user')
   user.update({
     DbUser.username: request.username,
     DbUser.email: request.email,
     DbUser.password: Hash.bcrypt(request.password)
   })
   db.commit()
-  return 'ok'
+  return 'Success'
 
-def delete_user(db: Session, id: int):
+def delete_user(db: Session, id: int,current_user):
   user = db.query(DbUser).filter(DbUser.id == id).first()
   if not user:
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
       detail=f'User with id {id} not found')
+  if current_user.id != id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='You are not allowed to delete this user')
   db.delete(user)
   db.commit()
-  return 'ok'
+  return 'Success'
